backend/requirements.txt
fastapi==0.95.2

uvicorn==0.22.0

python-dotenv==1.0.0

langchain==0.0.340

langchain-google-genai==0.0.4
google-generativeai==0.3.1
pydantic==1.10.12
typing-extensions==4.5.0